started correcting GCC warnings

+++ far/cmdline.cpp
@@ -329,7 +329,7 @@ int CommandLine::ProcessKey(int Key)
                KEY_END,        KEY_NUMPAD1
         };
 
-        for (int I=0;I< countof(UnmarkKeys);I++)
+        for (size_t I=0;I< countof(UnmarkKeys);I++)
           if (Key==UnmarkKeys[I])
           {
             CmdStr.Select(-1,0);
@@ -487,7 +487,7 @@ void CommandLine::GetPrompt(string &strDestStr)
       if (*Format==L'$')
       {
         wchar_t Chr=Upper(*++Format);
-        int I;
+        size_t I;
         for(I=0; I < countof(ChrFmt); ++I)
         {
           if(ChrFmt[I][0] == Chr)
